Moving api fetching logic to separate files
